live-agent/requirements.txt
# Core dependencies for the news agent system
google-generativeai>=0.8.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Optional dependencies for enhanced functionality
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
